Start with the preliminary proposal!
Thanks @yyhzuang for the reminder!

This commit adds to the preliminary proposal most things that were
in the full proposal's summary previously. I also moved the acronym
file over. (We'll reuse/link it if we write a full proposal, and
for the time being, the full proposal need not be buildable.)